Update docs and refactor trackedElement (now trackedObject)
